many changes, it's still in development;-)

+++ devices/simulation/TempSensorSim.py
@@ -1,4 +1,5 @@
 from base.Config import Config
+import os.path
 
 class TempSensor():
 
@@ -12,11 +13,11 @@ class TempSensor():
         pass
 
     def tick(self,pTimeDelta):
+        if (os.path.exists(self.confHolder.conf["simulation"]["temperature_sensor_file_name"])):
+            file = open(self.confHolder.conf["simulation"]["temperature_sensor_file_name"],"w")
 
-        file = open(self.confHolder.conf["simulation"]["temperature_sensor_file_name"],"w")
-
-        self._x_ = self._x_ + pTimeDelta * 0.25
-        file.write("33 00 4b 46 ff ff 02 10 f4 : crc=f4 YES\n33 00 4b 46 ff ff 02 10 f4 t=%05d" % (self._x_))
-        file.close()
-        print("##################### %05d" % self._x_)
+            self._x_ = self._x_ + pTimeDelta * 0.25
+            file.write("33 00 4b 46 ff ff 02 10 f4 : crc=f4 YES\n33 00 4b 46 ff ff 02 10 f4 t=%05d" % (self._x_))
+            file.close()
+            print("##################### %05d" % self._x_)
     pass
